requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.7.0
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
outcome==1.3.0.post0
packaging==24.2
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyDispatcher==2.0.7
pymongo==4.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.0.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
selenium==4.27.1
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
tldextract==5.1.3
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2